testing length + move tests
