+++ test/unit/test_disk_tensor.py
@@ -1,8 +1,8 @@
-import os, pathlib, tempfile, unittest, tarfile, io
+import pathlib, tempfile, unittest
 import numpy as np
 from tinygrad import Tensor, Device, dtypes
 from tinygrad.dtype import DType
-from tinygrad.nn.state import safe_load, safe_save, get_state_dict, torch_load, tar_extract
+from tinygrad.nn.state import safe_load, safe_save, get_state_dict, torch_load
 from tinygrad.helpers import Timing, fetch, temp, CI
 from tinygrad.device import is_dtype_supported
 
@@ -333,79 +333,6 @@ class TestDiskTensor(unittest.TestCase):
       on_dev = t.to(Device.DEFAULT).realize()
       np.testing.assert_equal(on_dev.numpy(), t.numpy())
 
-class TestTarExtractPAX(unittest.TestCase):
-  tar_format = tarfile.PAX_FORMAT
-  max_link_len = 1000_000
-
-  def setUp(self):
-    self.tar_path = pathlib.Path(tempfile.mktemp(".tar"))
-    self.test_files = {
-      'a/file1.txt': b'Hello, World!',
-      'a/b/file2.bin': b'\x00\x01\x02\x03\x04',
-      'empty_file.txt': b'',
-      'very'*15 + "/" + "very"*15 + '_long_filename.txt': b'Hello, World!'
-    }
-    self.test_dirs = ["a","a/b"]
-
-    with tarfile.open(self.tar_path, 'w', format=self.tar_format) as tar:
-      for dirname in self.test_dirs:
-        dir_info = tarfile.TarInfo(name=dirname)
-        dir_info.type = tarfile.DIRTYPE
-        tar.addfile(dir_info)
-
-      for filename, content in self.test_files.items():
-        file_info = tarfile.TarInfo(name=filename)
-        file_info.size = len(content)
-        tar.addfile(file_info, io.BytesIO(content))
-
-        if len(filename) < self.max_link_len:
-          link_info = tarfile.TarInfo(name=filename + ".lnk")
-          link_info.type = tarfile.SYMTYPE
-          link_info.linkname = filename
-          tar.addfile(link_info)
-
-  def tearDown(self): os.remove(self.tar_path)
-
-  def test_tar_extract_returns_dict(self):
-    result = tar_extract(Tensor(self.tar_path))
-    self.assertIsInstance(result, dict)
-
-  def test_tar_extract_correct_keys(self):
-    result = tar_extract(Tensor(self.tar_path))
-    self.assertEqual(set(result.keys()), set(self.test_files.keys()))
-
-  def test_tar_extract_content_size(self):
-    result = tar_extract(Tensor(self.tar_path))
-    for filename, content in self.test_files.items():
-      self.assertEqual(len(result[filename]), len(content))
-
-  def test_tar_extract_content_values(self):
-    result = tar_extract(Tensor(self.tar_path))
-    for filename, content in self.test_files.items():
-      np.testing.assert_array_equal(result[filename].numpy(), np.frombuffer(content, dtype=np.uint8))
-
-  def test_tar_extract_empty_file(self):
-    result = tar_extract(Tensor(self.tar_path))
-    self.assertEqual(len(result['empty_file.txt']), 0)
-
-  def test_tar_extract_non_existent_file(self):
-    with self.assertRaises(FileNotFoundError):
-      tar_extract(Tensor(pathlib.Path('non_existent_file.tar')))
-
-  def test_tar_extract_invalid_file(self):
-    with self.assertRaises(tarfile.ReadError):
-      tar_extract(Tensor(b'This is not a valid tar file'))
-
-  def test_tar_extract_invalid_file_long(self):
-    with self.assertRaises(tarfile.HeaderError):
-      tar_extract(Tensor(b'This is not a valid tar file'*100))
-
-class TestTarExtractUSTAR(TestTarExtractPAX):
-  tar_format = tarfile.USTAR_FORMAT
-  max_link_len = 100
-
-class TestTarExtractGNU(TestTarExtractPAX):
-  tar_format = tarfile.GNU_FORMAT
 
 class TestPathTensor(unittest.TestCase):
   def setUp(self):
